requirements.txt
fastapi
uvicorn[standard]

python-dotenv

sqlalchemy[asyncio]
asyncpg

pydantic

langchain
langchain-core
langchain-openai
aiogram
openai
aiogram-httpx